+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changes in Mini-XML 3.0
 
 - Changed the license to Apache 2.0 with exceptions (Issue #239)
-- All of the internal node structures are now moved out of the public header.
+- All of the internal node structures are now moved out of the public header
+  (Issue #240)
 - Fixed a potential buffer overflow when writing floating point data
   (Issue #233)
 - Moved `mxmldoc` to a new `codedoc` project whose focus is on generating