+++ CHANGES.md
@@ -11,6 +11,7 @@
 - The `mxmlSetCDATA`, `mxmlSetElement`, `mxmlSetOpaque`, `mxmlSetOpaquef`,
   `mxmlSetText`, and `mxmlSetTextf` functions caused a use-after-free bug if
   the value came from the same node (Issue #241)
+- The `mxmlSetOpaquef` and `mxmlSetTextf` functions did not work (Issue #244)
 
 
 # Changes in Mini-XML 2.12